restructure code in packages
