+++ src/main/kotlin/Laboratory_Work_No_3/Task1/Queue.kt
@@ -1,4 +1,4 @@
-package Laboratory_Work_No_3
+package Laboratory_Work_No_3.Task1
 
 interface Queue<T> {
     fun enqueue(node: T)
