--- src/main/kotlin/Laboratory_Work_No_3/main.kt
@@ -1,2 +1,0 @@
-package Laboratory_Work_No_3
-
